basic_fastapi/009-simple-database/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
psycopg2==2.9.9
psycopg2-binary==2.9.7
sqlalchemy==1.4.49
python-dotenv